fix file structure and components styling - separated the banner and search bar from a single component - fix header button styling - fix file structure in accordance with the team
