+++ src/components/Footer/Footer.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Box, Typography, Link } from '@mui/material';
 
-function Copyright() {
+const Copyright = () => {
   return (
     <Typography variant='body2' color='text.secondary' align='center'>
       {'Copyright © '}
@@ -12,9 +12,9 @@ function Copyright() {
       {'.'}
     </Typography>
   );
-}
+};
 
-const HomeFooter = () => {
+const Footer = () => {
   return (
     <Box sx={{ bgcolor: 'background.paper', p: 6 }} component='footer'>
       <Typography variant='h6' align='center' gutterBottom>
@@ -28,4 +28,4 @@ const HomeFooter = () => {
   );
 };
 
-export default HomeFooter;
+export default Footer;
