+++ src/components/Home/HomeBanner/HomeBanner.tsx
@@ -1,0 +1,19 @@
+import React from 'react';
+import { Stack } from '@mui/material';
+
+import HomeBannerImage from '@/assets/landscapeImage.jpg';
+
+const HomeBanner = () => {
+  return (
+    <Stack
+      direction='row'
+      justifyContent='end'
+      sx={{
+        pt: 24,
+        backgroundImage: `url(${HomeBannerImage})`,
+      }}
+    ></Stack>
+  );
+};
+
+export default HomeBanner;
