+++ src/components/Home/HomeRooms/HomeRoomList.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { Container, Grid } from '@mui/material';
 import HomeRoomItem from './HomeRoomItem';
 
 const DUMMY_ROOMS = [
@@ -10,11 +11,13 @@ const DUMMY_ROOMS = [
 
 const HomeRoomList = () => {
   return (
-    <>
-      {DUMMY_ROOMS.map(room => (
-        <HomeRoomItem key={room.id} id={room.id} roomName={room.roomName} vacancy={room.vacancy} bookmark={room.bookmark} />
-      ))}
-    </>
+    <Container sx={{ py: 8 }} maxWidth='md'>
+      <Grid container spacing={4}>
+        {DUMMY_ROOMS.map(room => (
+          <HomeRoomItem key={room.id} id={room.id} roomName={room.roomName} vacancy={room.vacancy} bookmark={room.bookmark} />
+        ))}
+      </Grid>
+    </Container>
   );
 };
